# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legazpi Souvenir Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeUser"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"

    post:
      operationId: addToCart
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"

    delete:
      operationId: clearCart
      tags: [cart]
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"

    delete:
      operationId: removeCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      parameters:
        - name: vendorId
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/{productId}/restock:
    post:
      operationId: restockProduct
      tags: [inventory]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestockRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/logs:
    get:
      operationId: listInventoryLogs
      tags: [inventory]
      parameters:
        - name: productId
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryLog"
...
  /kyc/submit:
    post:
      operationId: submitKyc
      tags: [kyc]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmitRequest"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycVerification"
  /kyc/status:
    get:
      operationId: getKycStatus
      tags: [kyc]
      responses:
        "200":
          description: KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycVerification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /kyc/{id}/review:
    put:
      operationId: reviewKyc
      tags: [kyc]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycReviewRequest"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycVerification"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SafeUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, staff, customer]
        isVerified:
          type: boolean
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, email, name, role, isVerified, kycStatus, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/SafeUser"
        token:
          type: string
      required: [user, token]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
      required: [email, password, name]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
...
    Product:
      type: object
      properties:
        id:
          type: string
        vendorId:
          type: string
        categoryId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        unit:
          type: string
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        isBestSeller:
          type: boolean
        isSeasonal:
          type: boolean
        isActive:
          type: boolean
        sku:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        ingredients:
          type: string
          nullable: true
        expirationMonths:
          type: integer
          nullable: true
        weight:
          type: string
          nullable: true
        vendorName:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, vendorId, categoryId, name, description, price, unit, stock, images, isBestSeller, isSeasonal, isActive, tags, rating, reviewCount, createdAt]
...
    CartItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
      required: [id, productId, quantity, product]
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        totalItems:
          type: integer
      required: [items, subtotal, totalItems]
    AddToCartRequest:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
      required: [productId, quantity]
    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: integer
      required: [quantity]
...
    InventoryItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        stock:
          type: integer
        sku:
          type: string
          nullable: true
        isActive:
          type: boolean
        price:
          type: number
      required: [productId, productName, vendorId, vendorName, stock, isActive, price]
    RestockRequest:
      type: object
      properties:
        quantity:
          type: integer
        notes:
          type: string
          nullable: true
      required: [quantity]
    InventoryLog:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        vendorId:
          type: string
        userId:
          type: string
        action:
          type: string
        quantity:
          type: integer
        previousStock:
          type: integer
        newStock:
          type: integer
        notes:
          type: string
          nullable: true
        productName:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, productId, vendorId, userId, action, quantity, previousStock, newStock, createdAt]
...
    KycVerification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
        address:
          type: string
        idType:
          type: string
        idNumber:
          type: string
        reviewNotes:
          type: string
          nullable: true
        submittedAt:
          type: string
        reviewedAt:
          type: string
          nullable: true
      required: [id, userId, status, firstName, lastName, birthDate, address, idType, idNumber, submittedAt]
    KycSubmitRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
        address:
          type: string
        idType:
          type: string
        idNumber:
          type: string
      required: [firstName, lastName, birthDate, address, idType, idNumber]
    KycReviewRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        reviewNotes:
          type: string
          nullable: true
      required: [status]