# lib/api-zod/src/generated/types/kycReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legazpi Souvenir Market API
 * OpenAPI spec version: 0.1.0
 */
import type { KycReviewRequestStatus } from "./kycReviewRequestStatus";

export interface KycReviewRequest {
  status: KycReviewRequestStatus;
  reviewNotes?: string | null;
}